+++ clauses/make/make.py
@@ -1,37 +1,29 @@
 from ast import literal_eval
 from re import search
-from typing import Optional, Union
+from typing import Union
 
-from pydantic import BaseModel
-
-from enums.statement import Statement
+from models.enums.statement import Statement
 from exceptions.statements import IllegalParameterType, MissingNodeLabel, StatementError
+from models.statement import ParsedStatement
 
 
-class ParsedStatement(BaseModel):
-    clause: Statement
-    handle: Optional[str]
-    node_label: str
-    params: Optional[dict]
-
-
-def parse_make_statment(input_string: str) -> ParsedStatement:
+def parse_make_statment(statement_string: str) -> ParsedStatement:
 
     # validate that the correct statement is called
 
     make_statement_pattern = r"""(?P<clause>MAKE|make)"""
 
-    if not search(make_statement_pattern, input_string):
-        raise StatementError(input_string)
+    if not search(make_statement_pattern, statement_string):
+        raise StatementError(statement_string)
 
     node_label_pattern = r""":\s*(?P<node_label>\w+)"""
 
-    if not search(node_label_pattern, input_string):
-        raise MissingNodeLabel(input_string)
+    if not search(node_label_pattern, statement_string):
+        raise MissingNodeLabel(statement_string)
 
-    parsing_pattern = r"""(?P<clause>MAKE|make)\s*\((?P<handle>\w+)?\s*:\s*(?P<node_label>\w+)\s*{?(?P<params>[\w\'\:\|\s\-\.\,\[\]]+)?\s*}?\s*\)"""
+    parsing_pattern = rf"""{make_statement_pattern}\s*\((?P<handle>\w+)?\s*{node_label_pattern}\s*{{?(?P<params>[\w\'\:\|\s\-\.\,\[\]]+)?\s*}}?\s*\)"""
 
-    if matches := search(parsing_pattern, input_string):
+    if matches := search(parsing_pattern, statement_string):
 
         handle = matches.group("handle")
         node_label = matches.group("node_label")
@@ -44,7 +36,11 @@ def parse_make_statment(input_string: str) -> ParsedStatement:
             params = None
 
         return ParsedStatement(
-            clause=Statement.MAKE, handle=handle, node_label=node_label, params=params
+            clause=Statement.MAKE,
+            handle=handle,
+            node_label=node_label,
+            params=params,
+            statement_string=statement_string,
         )
 
 
@@ -66,13 +62,14 @@ def build_params_from_string(params_string: str) -> dict:
             if value.strip() == "":
                 raise IllegalParameterType(param)
 
-            value_parsed = value_to_correct_data_type(value)
+            value_parsed = string_to_correct_data_type(value)
+
             params_dict[key] = value_parsed
 
     return params_dict
 
 
-def value_to_correct_data_type(value: str) -> Union[float, int, str, list]:
+def string_to_correct_data_type(value: str) -> Union[float, int, str, list]:
 
     value = value.strip('"')
     value = value.strip("'")
